Add function (*Decoder).More() and (*Decoder).SetMaxStringLength() Change the integer type of depth, maxDepth etc in Decoder and Encoder

+++ vbs/error.go
@@ -16,7 +16,7 @@ func (e *UnsupportedTypeError) Error() string {
 
 
 type DepthOverflowError struct {
-	MaxDepth int
+	MaxDepth int16
 }
 
 func (e *DepthOverflowError) Error() string {
